+++ src/parallelization/stripingsimulator.h
@@ -208,7 +208,7 @@ private:
             validateLoads(newWorkloads, oldWorkloads);
             newPartitionsSendBuffer = workloadsToPartitions(newWorkloads);
 
-            for (size_t i = 0; i < mpilayer.size(); i++) {
+            for (int i = 0; i < mpilayer.size(); i++) {
                 mpilayer.sendVec(&newPartitionsSendBuffer, i, BALANCELOADS);
             }
         }
